feat(ui): fix build, rename AppImage to Image
