+++ apps/dsb-client-gateway-frontend/src/components/Sidebar/Sidebar.styles.ts
@@ -1,5 +1,4 @@
 import { makeStyles } from 'tss-react/mui';
-import { alpha } from '@mui/material/styles';
 
 const drawerWidth = 264;
