+++ libs/dsb-client-gateway/ui/channels/src/lib/components/ChannelImage/ChannelImage.styles.ts
@@ -4,7 +4,6 @@ export const useStyles = makeStyles()((theme) => ({
   imageWrapper: {
     borderRadius: 6,
     background: theme.palette.background.default,
-    marginBottom: 4,
     display: 'flex',
     alignItems: 'center',
     justifyContent: 'center',
