+++ libs/dsb-client-gateway/ui/channels/src/lib/containers/modals/Details/Details.tsx
@@ -31,7 +31,14 @@ export const Details: FC = () => {
         {channel && (
           <>
             <Box className={classes.channelWrapper}>
-              <ChannelImage type={channel.type} wrapperProps={{ width: '62px', height: '52px' }} />
+              <ChannelImage
+                type={channel.type}
+                wrapperProps={{
+                  width: '62px',
+                  height: '58px',
+                  marginBottom: '4px',
+                }}
+              />
               <Typography className={classes.type}>
                 {getChannelType(channel.type)}
               </Typography>
