--- libs/dsb-client-gateway/ui/core/src/components/common/AppImage/index.ts
@@ -1,1 +1,0 @@
-export * from './AppImage';
