+++ libs/dsb-client-gateway/ui/core/src/components/common/Image/Image.styles.ts
