+++ libs/dsb-client-gateway/ui/core/src/components/common/Image/index.ts
@@ -1,0 +1,1 @@
+export * from './Image';
