+++ libs/dsb-client-gateway/ui/core/src/components/common/index.ts
@@ -1,3 +1,3 @@
 export * from './Swal';
-export * from './AppImage';
+export * from './Image';
 export * from './Backdrop';
