+++ libs/dsb-client-gateway/ui/topics/src/lib/components/ApplicationInfo/ApplicationInfo.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { Typography, Box } from '@mui/material';
 import { ApplicationDTO } from '@dsb-client-gateway/dsb-client-gateway-api-client';
-import { AppImage } from '@dsb-client-gateway/ui/core';
+import { Image } from '@dsb-client-gateway/ui/core';
 import { useStyles } from './ApplicationInfo.styles';
 
 interface ApplicationInfoProps {
@@ -12,7 +12,7 @@ export const ApplicationInfo: FC<ApplicationInfoProps> = ({ application }) => {
   const { classes } = useStyles();
   return (
     <Box>
-      <AppImage src={application.logoUrl} className={classes.appImage} />
+      <Image src={application.logoUrl} className={classes.appImage} />
       <Box mt={2.5}>
         <Typography variant="body2" className={classes.label}>
           Application name
